refactor new prog model structure v1
