+++ azure/functions/__init__.py
@@ -16,9 +16,6 @@ from ._durable_functions import OrchestrationContext, EntityContext
 from .meta import get_binding_registry
 from .extension import (ExtensionMeta, FunctionExtensionException,
                         FuncExtensionBase, AppExtensionBase)
-from ._decorators import DataType
-from .decorators import AuthLevel, BlobInput, BlobOutput, EventHubTrigger, \
-    FunctionsApp, Http, HttpMethod, HttpTrigger
 
 # Import binding implementations to register them
 from . import blob  # NoQA
