+++ azure/functions/decorator/__init__.py
@@ -1,0 +1,6 @@
+from .core import FunctionsApp
+
+
+__all__ = [
+    'FunctionsApp'
+]
